requirements.txt
numpy
pandas
scipy
matplotlib
pycuda
tensorflow
pyyaml
statsmodels